minorProj/requirements.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
lightgbm>=4.3.0
xgboost>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
joblib>=1.3.2
matplotlib>=3.8.3
seaborn>=0.13.2
plotly>=5.19.0
pytest>=8.0.0
python-dotenv>=1.0.1
pyyaml>=6.0.1